pyproject.toml
[tool.poetry]
name = "aibot-story-builder"
version = "0.1.0"
description = "Turn an image or text into an audio story"
authors = ["alex00pep <alexmtnezf@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [ 
{ include = "app", from = "src" }
]
[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.26.0"
python-dotenv = "^1.0.0"
transformers = "^4.33.1"
torch = "^2.0.1"
langchain = "^0.0.291"
einops = "^0.6.1"
accelerate = "^0.23.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
